+++ Changelog.md
@@ -3,6 +3,7 @@
 * [#3](https://github.com/vizor-games/jac/issues/3) Allow reference local values inside string evaluations
 * [#4](https://github.com/vizor-games/jac/issues/4) Add implicit `^top` and `^base` profiles
 * Allow reference local values inside string evaluations at top level keys
+* Allow reference profile name at config evaluation
 
 # Version 0.0.3